requirements.txt
Django==4.1.3
djangorestframework==3.14.0
requests==2.28.1
responses==0.22.0
pytest-cov==4.0.0
pytest-django==4.5.2
